+++ IzvorniKod/backend/src/main/java/vidoje/eventko/controller/UserController.java
@@ -1,6 +1,7 @@
 package vidoje.eventko.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import vidoje.eventko.domain.User;
@@ -11,6 +12,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/user")
@@ -19,30 +21,60 @@ public class UserController {
     private UserService userService;
     @GetMapping("/users")
     public ResponseEntity<UserListResponseDTO> listUsers(HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new UserListResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna", null), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
 
-        return  ResponseEntity.ok(new UserListResponseDTO(userService.listAllNotBlocked(userId)));
+        return  ResponseEntity.ok(new UserListResponseDTO("", userService.listAllNotBlocked(userId)));
     }
 
     @GetMapping("")
     public ResponseEntity<UserInfoResponseDTO> userInfo(HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new UserInfoResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna", null), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
 
-        return ResponseEntity.ok(new UserInfoResponseDTO(userService.getUserById(userId)));
+        return ResponseEntity.ok(new UserInfoResponseDTO("", userService.getUserById(userId)));
     }
 
     @PostMapping("/logout")
     public ResponseEntity<MessageResponseDTO> destroySession(HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
         request.getSession().invalidate();
         return ResponseEntity.ok(new MessageResponseDTO("Uspješna Odjava"));
     }
 
     @PostMapping("/friend")
     public ResponseEntity<MessageResponseDTO> addFriend(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
         User user = userService.getUserById(userId);
 
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
         User other = userService.getUserById(dto.getUserId());
+
+        if(user.getBlockedBy().contains(other)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik je blokirao trenutnog"), HttpStatus.BAD_REQUEST);
+        }
+
+        if(other.getFriends().contains(user)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik je već prijatelj trenutnog"), HttpStatus.BAD_REQUEST);
+
+        }
+
         Set<User> userFriends = user.getFriends();
         userFriends.add(other);
         user.setFriends(userFriends);
@@ -56,10 +88,24 @@ public class UserController {
 
     @PostMapping("/unfriend")
     public ResponseEntity<MessageResponseDTO> removeFriend(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
         User user = userService.getUserById(userId);
 
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
         User other = userService.getUserById(dto.getUserId());
+
+        if(!other.getFriends().contains(user)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik već nije prijatelj trenutnog"), HttpStatus.BAD_REQUEST);
+
+        }
+
         Set<User> userFriends = user.getFriends();
         userFriends.remove(other);
         user.setFriends(userFriends);
@@ -73,13 +119,35 @@ public class UserController {
 
     @PostMapping("/block")
     public ResponseEntity<MessageResponseDTO> blockUser(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
         User user = userService.getUserById(userId);
 
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
         User other = userService.getUserById(dto.getUserId());
+
+        if(other.getBlockedBy().contains(user)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik je već blokirao drugog"), HttpStatus.BAD_REQUEST);
+        }
+
+
         Set<User> blockedBy = other.getBlockedBy();
         blockedBy.add(user);
         other.setBlockedBy(blockedBy);
+
+        Set<User> userFriends = user.getFriends();
+        userFriends.remove(other);
+        user.setFriends(userFriends);
+
+        userFriends = other.getFriends();
+        userFriends.remove(user);
+        other.setFriends(userFriends);
 
 
         return ResponseEntity.ok(new MessageResponseDTO("Korisnik je blokiran"));
@@ -87,6 +155,21 @@ public class UserController {
 
     @PostMapping("/suspend")
     public ResponseEntity<MessageResponseDTO> suspendUser(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
+        Long userId = (Long) request.getSession().getAttribute("USER_ID");
+        User user = userService.getUserById(userId);
+
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
+        if(!user.getRoles().stream().map(r -> r.getId()).collect(Collectors.toSet()).contains(3)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Samo korisnik s ulogom moderator može suspendirati druge"), HttpStatus.BAD_REQUEST);
+        }
+
 
         User other = userService.getUserById(dto.getUserId());
         other.setSuspended(true);
@@ -97,6 +180,20 @@ public class UserController {
 
     @PostMapping("/unsuspend")
     public ResponseEntity<MessageResponseDTO> unsuspendUser(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
+        Long userId = (Long) request.getSession().getAttribute("USER_ID");
+        User user = userService.getUserById(userId);
+
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
+        if(!user.getRoles().stream().map(r -> r.getId()).collect(Collectors.toSet()).contains(3)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Samo korisnik s ulogom moderator može odsuspendirati druge"), HttpStatus.BAD_REQUEST);
+        }
 
         User other = userService.getUserById(dto.getUserId());
         other.setSuspended(false);
@@ -107,6 +204,20 @@ public class UserController {
 
     @PostMapping("/delete")
     public ResponseEntity<MessageResponseDTO> removeUser(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new MessageResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna"), HttpStatus.BAD_REQUEST);
+        }
+
+        Long userId = (Long) request.getSession().getAttribute("USER_ID");
+        User user = userService.getUserById(userId);
+
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new MessageResponseDTO("Drugi korisnik s tim ID-jem ne postoji"), HttpStatus.BAD_REQUEST);
+        }
+
+        if(!user.getRoles().stream().map(r -> r.getId()).collect(Collectors.toSet()).contains(4)) {
+            return new ResponseEntity<>(new MessageResponseDTO("Samo korisnik s ulogom admin može izbrisati druge"), HttpStatus.BAD_REQUEST);
+        }
 
         userService.delete(dto.getUserId());
 
@@ -114,12 +225,19 @@ public class UserController {
     }
 
 
-
     @GetMapping("/friends")
     public ResponseEntity<UserListResponseDTO> viewFriends(@Valid @RequestBody UserRequestDTO dto, HttpServletRequest request) {
+        if(!request.isRequestedSessionIdValid()) {
+            return new ResponseEntity<>(new UserListResponseDTO("Korisnik nije ulogiran i/ili FE-BE sesija nije aktivna", null), HttpStatus.BAD_REQUEST);
+        }
+
         Long userId = (Long) request.getSession().getAttribute("USER_ID");
         User user = userService.getUserById(userId);
 
-        return ResponseEntity.ok(new UserListResponseDTO(user.getFriends().stream().toList()));
+        if(!userService.exists(dto.getUserId())) {
+            return new ResponseEntity<>(new UserListResponseDTO("Drugi korisnik s tim ID-jem ne postoji", null), HttpStatus.BAD_REQUEST);
+        }
+
+        return ResponseEntity.ok(new UserListResponseDTO("", user.getFriends().stream().toList()));
     }
 }
